rename `ChatInstructionsFileLocator` utility and move it under `/common` to be used by more code
